Rewrite `calculateInsertPosition` getter as action
