+++ client/store/modules/repository/activities/actions.js
@@ -1,6 +1,7 @@
+import { getDescendants as getDeepChildren, getOutlineChildren } from 'utils/activity';
 import calculatePosition from 'utils/calculatePosition';
+import findIndex from 'lodash/findIndex';
 import generateActions from '@/store/helpers/actions';
-import { getDescendants as getDeepChildren } from 'utils/activity';
 import request from '@/api/request';
 
 const { api, fetch, get, reset, save, setEndpoint, update } = generateActions();
@@ -35,7 +36,18 @@ const clone = ({ commit }, mapping) => {
     .then(({ data: { data } }) => commit('fetch', api.processEntries(data)));
 };
 
+const calculateInsertPosition = ({ state }, { activity, anchor, action }) => {
+  const items = getOutlineChildren(state.items, activity.parentId);
+  const newPosition = anchor ? findIndex(items, { id: anchor.id }) : 1;
+  const isFirstChild = !anchor ||
+    (activity.parentId !== anchor.parentId) ||
+    (newPosition === -1);
+  const context = { items, newPosition, isFirstChild, action };
+  return calculatePosition(context);
+};
+
 export {
+  calculateInsertPosition,
   clone,
   get,
   fetch,
