+++ client/store/modules/repository/activities/getters.js
@@ -1,11 +1,8 @@
 import {
   getDescendants as getDeepChildren,
-  getOutlineChildren,
   getAncestors as getParents
 } from 'utils/activity';
-import calculatePosition from 'utils/calculatePosition';
 import find from 'lodash/find';
-import findIndex from 'lodash/findIndex';
 
 export const activities = state => state.items;
 
@@ -29,17 +26,5 @@ export const getLineage = state => {
     const ancestors = getParents(state.items, activity);
     const descendants = getDeepChildren(state.items, activity);
     return [...ancestors, ...descendants];
-  };
-};
-
-export const calculateInsertPosition = state => {
-  return (activity, anchor, action) => {
-    const items = getOutlineChildren(state.items, activity.parentId);
-    const newPosition = anchor ? findIndex(items, { id: anchor.id }) : 1;
-    const isFirstChild = !anchor ||
-      (activity.parentId !== anchor.parentId) ||
-      (newPosition === -1);
-    const context = { items, newPosition, isFirstChild, action };
-    return calculatePosition(context);
   };
 };
